Clean up testing tasks
