+++ rakelib/test_puppet.rake
@@ -1,9 +1,9 @@
-require 'simp/packer/tests/puppet'
+require 'simp/tests/puppet'
 
 namespace :test do
   desc 'Test Puppet modules'
   task :puppet do
-    include Simp::Packer::Tests::Puppet
+    include Simp::Tests::Puppet
     Dir[File.join('puppet', 'modules', '*')].each do |dir|
       Dir.chdir dir do
         run_puppet_rake_tests
